riscv: fix tmp register conflict when loading a large immediate
`tmp_r` can be the same as base register, so we cannot use it as a tmp register
for `load_immediate()`. This was discovered while running pcre2 tests on QEMU
with Sv48 kernel. In that case, sljit will move a relatively large heap memory
address immediate to the tmp register, so `load_immediate` actually uses tmp_r,
causing a register conflict.

Reported-by: Felix Yan <felixonmars@archlinux.org>
Reported-by: dramforever <dramforever@live.com>